Avoid camel_stream_printf().
camel_stream_printf() is next on the chopping block.

Use g_strdup_printf() or a GString to construct a formatted string in
memory, pass it to camel_stream_write() in one go, and then check for
errors (unless it's a memory stream).

This commit also fixes some obvious memory leaks in the immediate
vicinity.